Use specialized faces for weaver keywords
Collector keywords have a command you actually write, and another
command they're equivalent to, as the weaving process will transform
the former into the latter.

This change will make the collector commands be highlighted with the
face of their replacement command, if we have a special face for that
available (which for now means head[1234] only).