Avoid changing buffers
When (re)building its list of modules, perldoc.el uses a temporary
buffer to collect its results, and consecutively doesn't clean it
up. That causes the current buffer to be changed when invoking
anything that triggers module list building interactively, and results
in the consecutive insert call operating on the wrong buffer.

This works around what I think is a bug in perldoc.el.